requirements.txt
streamlit==1.0.0
tensorflow-cpu==2.6.0
Pillow==8.4.0
